+++ .babelrc
@@ -1,4 +1,4 @@
 {
-  "presets": [ "es2015" ],
+  "presets": [ "es2015"],
   "plugins": ["transform-object-rest-spread"]
 }
