+++ lib/check.js
@@ -1,14 +1,12 @@
 import { is } from './util/type'
 import { info, assert } from './util/warn'
 import base64url from 'base64url'
-
+import {dict, httpRequest} from './util/http'
 let _appId,// 唯一标识
   _timestamp,// 生成签名的时间戳
   _nonceStr,// 生成签名的随机串
   _signature,// 签名
-  _event,
-  _url,
-  _dict
+  _url
 
 /**
  * 判断一个字符串是否为空，如果非字符串参数，则抛出类型错误异常
@@ -27,44 +25,50 @@ function checkSign() {
     if (!_strIsEmpty(_appId) && !_strIsEmpty(_timestamp) && !_strIsEmpty(_nonceStr) && !_strIsEmpty(_signature)) {
       window.getCheckResult = (res) => {
         const temp = JSON.parse(res)
-        if (temp[_dict.statusCodeKey] === _dict.statusCodeSuccess) {
+        if (temp[dict.statusCodeKey] === dict.statusCodeSuccess) {
           info('验签完成')
           resolve(temp)
         } else {
-          let errMsg = base64url.decode(temp[_dict.msgKey])
+          const errMsg = base64url.decode(temp[dict.msgKey])
           assert(errMsg)
           reject(new Error(errMsg))
         }
       }
-      const command = {
-        event: 'AjaxEvent',
-        action: 'sendOriginalRequest',
-        listener: 'getCheckResult',
-        params: {
-          transcode: 'ThridJsAccessControl.do',
-          params: {
-            appId: _appId,
-            timestamp: _timestamp,
-            nonceStr: _nonceStr,
-            signature: _signature,
-            url: _url
-          }
-        }
+      const params = {
+        appId: _appId,
+        timestamp: _timestamp,
+        nonceStr: _nonceStr,
+        signature: _signature,
+        url: _url
       }
-      _event(command)
+      httpRequest('ThridJsAccessControl.do', params, 'getCheckResult')
+      // const command = {
+      //   event: 'AjaxEvent',
+      //   action: 'sendOriginalRequest',
+      //   listener: 'getCheckResult',
+      //   params: {
+      //     transcode: 'ThridJsAccessControl.do',
+      //     params: {
+      //       appId: _appId,
+      //       timestamp: _timestamp,
+      //       nonceStr: _nonceStr,
+      //       signature: _signature,
+      //       url: _url
+      //     }
+      //   }
+      // }
+      // _event(command)
     } else {
       reject(new Error(`config err:请核对验签参数`))
     }
   })
 }
 
-export const install = function (eventFunc, appId, timestamp, nonceStr, signature, url, dict) {
-  _event = eventFunc
+export const install = function (eventFunc, appId, timestamp, nonceStr, signature, url) {
   _appId = appId
   _timestamp = timestamp
   _nonceStr = nonceStr
   _signature = signature
   _url = url
-  _dict = dict
   return checkSign()
 }
