+++ lib/index.js
@@ -5,15 +5,10 @@ import {install as navigation} from './navigation.js'
 // import {install as share} from './share.js'
 import {install as user} from './user.js'
 import device from './util/device'
-import MixinPlugin from "./util/mixin-plugin";
+import MixinPlugin from "./util/mixin-plugin"
+import {init as http} from './util/http'
 // 和业务相关，只需要判断一下几种类型
 const isRunNative = (device.isAndroid || device.isAndroidPad || device.isIpad || device.isIphone)
-
-const dict = {
-  statusCodeKey: 'ReturnCode',
-  statusCodeSuccess: '000000',
-  msgKey: 'ReturnMessage'
-}
 
 export class JSBridge {
   constructor() {
@@ -40,9 +35,11 @@ export class JSBridge {
       jsBridge
     })
 
+    http(eventFunc)
+
     return new Promise((resolve, reject) => {
-      check(eventFunc, appId, timestamp, nonceStr, signature, url, dict).then((res) => {
-        const plugins = [navigation(eventFunc, dict), user(eventFunc, res.level, dict )]
+      check(eventFunc, appId, timestamp, nonceStr, signature, url).then(() => {
+        const plugins = [navigation(eventFunc), user(eventFunc, appId)]
         MixinPlugin.mixin(jsBridge, plugins)
         resolve(jsBridge)
       }).catch(err => {
