+++ lib/user.js
@@ -1,6 +1,8 @@
-let _statusCodeSuccess, _event,
-  _ruleLevel /*权限等级：0:最高权限 1：（不包含身份证号和手机号） */
+import base64url from 'base64url'
+import {dict, httpRequest} from './util/http'
 
+let _statusCodeSuccess, _event,
+  _appId /* 第三方唯一识别嘛 */
 
 /**
  * 和导航相关的工具模块
@@ -13,23 +15,17 @@ const plugin = {
    */
   getUserInfo() {
     return new Promise((resolve, reject) => {
-      _event({
-        event: 'UserInfoManageEvent',
-        action: 'getUserInfo'
-      }).then(res => {
-        switch (_ruleLevel) {
-          case 0 :
-            resolve(res)
-            break
-          case 1 :
-            delete res.userInfo.UserIdType
-            delete res.userInfo.UserIdNo
-            delete res.userInfo.MobilePhoneEncrypt
-            resolve(res)
+      window.getUserInfoResult = (res) => {
+        window.getUserInfoResult = null
+        const temp = JSON.parse(res)
+        if (temp[dict.statusCodeKey] === dict.statusCodeSuccess) {
+          resolve(temp)
+        } else {
+          const errMsg = base64url.decode(temp[dict.msgKey])
+          reject(new Error(errMsg))
         }
-      }).catch((err) => {
-        reject(err)
-      })
+      }
+      httpRequest('GetUserInfoForSdk.do', {appId: _appId}, 'getUserInfoResult')
     })
   },
   /**
@@ -73,11 +69,11 @@ const plugin = {
   }
 }
 
-export const install = function (eventFunc, ruleLevel, {
+export const install = function (eventFunc, appId, {
   statusCodeSuccess
 }) {
   _statusCodeSuccess = statusCodeSuccess
   _event = eventFunc
-  _ruleLevel = typeof (ruleLevel) === 'undefined' ? 1 : ruleLevel
+  _appId = appId
   return plugin
 }
